// Hawa-front/src/app/shared/model/sign-up.model.ts
export class SignUp {
    userName: string;
    password: string;
    email: string;
    type: string;
    organizationName: string;
    taxNumber: string;
    acronymName: string;
    representative: string;
    phone: string;
    fax: string;
    website: string;
    stateProvinceID: number;
    districtID: number;
    communeID: number;
    houseNumber: string;
    address: string;
    identityCard: string;
};